Extract chart into ChartUiView
